+++ ichnaea/service/geolocate/tests.py
@@ -63,7 +63,9 @@ class TestGeolocate(AppTestCase):
             status=200)
 
         self.check_stats(
-            counter=[self.metric_url + '.200', self.metric + '.api_key.test']
+            counter=[self.metric_url + '.200',
+                     self.metric + '.api_key.test',
+                     self.metric + '.api_log.test.cell_hit']
         )
 
         self.assertEqual(res.content_type, 'application/json')
@@ -92,7 +94,8 @@ class TestGeolocate(AppTestCase):
                 ]},
             status=200)
         self.check_stats(
-            counter=[self.metric + '.api_key.test'])
+            counter=[self.metric + '.api_key.test',
+                     self.metric + '.api_log.test.wifi_hit'])
         self.assertEqual(res.content_type, 'application/json')
         self.assertEqual(res.json, {"location": {"lat": 1.001,
                                                  "lng": 1.002},
@@ -123,6 +126,7 @@ class TestGeolocate(AppTestCase):
         # Make sure to get two counters, a timer, and no traceback
         self.check_stats(
             counter=[self.metric + '.api_key.test',
+                     self.metric + '.api_log.test.wifi_miss',
                      self.metric_url + '.404'],
             timer=[self.metric_url],
         )