+++ ichnaea/service/geosubmit/views.py
@@ -212,7 +212,9 @@ def process_single(request):
         result = search_all_sources(
             session, 'geosubmit', data,
             client_addr=request.client_addr,
-            geoip_db=request.registry.geoip_db)
+            geoip_db=request.registry.geoip_db,
+            api_key_log=getattr(request, 'api_key_log', False),
+            api_key_name=getattr(request, 'api_key_name', None))
     else:
         result = {'lat': first_item['latitude'],
                   'lon': first_item['longitude'],